# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Test Results Dashboard API
  version: 0.1.0
  description: Read test runs, and upload JUnit XML reports from CI.
servers:
  - url: http://localhost:8080
paths:
...
  /api/runs/{id}:
    get:
      summary: One run with its test cases
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Run }
        "404": { description: Not found }
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }